requirements.txt
alpaca-py
pydantic>=2.0
python-telegram-bot==13.15
pyyaml
pytz
schedule
urllib3>=1.26,<2
